+++ src/api/flow.ts
@@ -2,9 +2,12 @@ import { action } from "./action"
 
 let generatorId = 0
 
-// non-decorator forms
-export function flow<R>(generator: () => IterableIterator<any>): () => Promise<R>
-export function flow<A1>(generator: (a1: A1) => IterableIterator<any>): (a1: A1) => Promise<any> // Ideally we want to have R instead of Any, but cannot specify R without specifying A1 etc... 'any' as result is better then not specifying request args
+export type CancellablePromise<T> = Promise<T> & { cancel(): void }
+
+export function flow<R>(generator: () => IterableIterator<any>): () => CancellablePromise<R>
+export function flow<A1>(
+    generator: (a1: A1) => IterableIterator<any>
+): (a1: A1) => CancellablePromise<any> // Ideally we want to have R instead of Any, but cannot specify R without specifying A1 etc... 'any' as result is better then not specifying request args
 export function flow<A1, A2, A3, A4, A5, A6, A7, A8>(
     generator: (
         a1: A1,
@@ -16,22 +19,22 @@ export function flow<A1, A2, A3, A4, A5, A6, A7, A8>(
         a7: A7,
         a8: A8
     ) => IterableIterator<any>
-): (a1: A1, a2: A2, a3: A3, a4: A4, a5: A5, a6: A6, a7: A7, a8: A8) => Promise<any>
+): (a1: A1, a2: A2, a3: A3, a4: A4, a5: A5, a6: A6, a7: A7, a8: A8) => CancellablePromise<any>
 export function flow<A1, A2, A3, A4, A5, A6, A7>(
     generator: (a1: A1, a2: A2, a3: A3, a4: A4, a5: A5, a6: A6, a7: A7) => IterableIterator<any>
-): (a1: A1, a2: A2, a3: A3, a4: A4, a5: A5, a6: A6, a7: A7) => Promise<any>
+): (a1: A1, a2: A2, a3: A3, a4: A4, a5: A5, a6: A6, a7: A7) => CancellablePromise<any>
 export function flow<A1, A2, A3, A4, A5, A6>(
     generator: (a1: A1, a2: A2, a3: A3, a4: A4, a5: A5, a6: A6) => IterableIterator<any>
-): (a1: A1, a2: A2, a3: A3, a4: A4, a5: A5, a6: A6) => Promise<any>
+): (a1: A1, a2: A2, a3: A3, a4: A4, a5: A5, a6: A6) => CancellablePromise<any>
 export function flow<A1, A2, A3, A4, A5>(
     generator: (a1: A1, a2: A2, a3: A3, a4: A4, a5: A5) => IterableIterator<any>
-): (a1: A1, a2: A2, a3: A3, a4: A4, a5: A5) => Promise<any>
+): (a1: A1, a2: A2, a3: A3, a4: A4, a5: A5) => CancellablePromise<any>
 export function flow<A1, A2, A3, A4>(
     generator: (a1: A1, a2: A2, a3: A3, a4: A4) => IterableIterator<any>
-): (a1: A1, a2: A2, a3: A3, a4: A4) => Promise<any>
+): (a1: A1, a2: A2, a3: A3, a4: A4) => CancellablePromise<any>
 export function flow<A1, A2, A3>(
     generator: (a1: A1, a2: A2, a3: A3) => IterableIterator<any>
-): (a1: A1, a2: A2, a3: A3) => Promise<any>
+): (a1: A1, a2: A2, a3: A3) => CancellablePromise<any>
 export function flow<A1, A2>(
     generator: (a1: A1, a2: A2) => IterableIterator<any>
 ): (a1: A1, a2: A2) => Promise<any>
@@ -45,13 +48,17 @@ export function flow(generator: Function) {
     return function() {
         const ctx = this
         const args = arguments
-        return new Promise(function(resolve, reject) {
-            const runId = ++generatorId
+        const runId = ++generatorId
+        const gen = action(`${name} - runid: ${runId} - init`, generator).apply(ctx, args)
+        let rejector: (error: any) => void
+        let pendingPromise: CancellablePromise<any> | undefined = undefined
+
+        const res = new Promise(function(resolve, reject) {
             let stepId = 0
-            const gen = action(`${name} - runid: ${runId} - init`, generator).apply(ctx, args)
-            onFulfilled(undefined) // kick off the process
+            rejector = reject
 
             function onFulfilled(res: any) {
+                pendingPromise = undefined
                 let ret
                 try {
                     ret = action(`${name} - runid: ${runId} - yield ${stepId++}`, gen.next).call(
@@ -61,11 +68,12 @@ export function flow(generator: Function) {
                 } catch (e) {
                     return reject(e)
                 }
+
                 next(ret)
-                return null
             }
 
             function onRejected(err: any) {
+                pendingPromise = undefined
                 let ret
                 try {
                     ret = action(`${name} - runid: ${runId} - yield ${stepId++}`, gen.throw).call(
@@ -79,12 +87,29 @@ export function flow(generator: Function) {
             }
 
             function next(ret: any) {
+                if (ret && typeof ret.then === "function") {
+                    // an async iterator
+                    ret.then(next, reject)
+                    return
+                }
                 if (ret.done) return resolve(ret.value)
-                // TODO: support more type of values? See https://github.com/tj/co/blob/249bbdc72da24ae44076afd716349d2089b31c4c/index.js#L100
-                if (!ret.value || typeof ret.value.then !== "function")
-                    fail("Only promises can be yielded to asyncAction, got: " + ret)
-                return ret.value.then(onFulfilled, onRejected)
+                pendingPromise = Promise.resolve(ret.value) as any
+                return pendingPromise!.then(onFulfilled, onRejected)
+            }
+
+            onFulfilled(undefined) // kick off the process
+        }) as any
+
+        res.cancel = action(`${name} - runid: ${runId} - cancel`, function() {
+            try {
+                if (pendingPromise && typeof pendingPromise.cancel === "function")
+                    pendingPromise.cancel()
+                gen.return()
+                rejector(new Error("FLOW_CANCELLED"))
+            } catch (e) {
+                rejector(e) // there could be a throwing finally block
             }
         })
+        return res
     }
 }