+++ test/base/flow.js
@@ -1,4 +1,5 @@
-var mobx = require("../../src/mobx.ts")
+import * as mobx from "../../src/mobx.ts"
+const { flow } = mobx
 
 function delay(time, value, shouldThrow = false) {
     return new Promise((resolve, reject) => {
@@ -157,3 +158,172 @@ function stripEvents(events) {
         return e
     })
 }
+
+test("flows can be cancelled - 1 - uncatched cancellation", done => {
+    let steps = 0
+    const start = flow(function*() {
+        steps = 1
+        yield Promise.resolve()
+        steps = 2
+    })
+
+    const promise = start()
+    promise.then(
+        () => {
+            debugger
+            fail()
+        },
+        err => {
+            debugger
+            expect(steps).toBe(1)
+            expect("" + err).toBe("Error: FLOW_CANCELLED")
+            done()
+        }
+    )
+    debugger
+    promise.cancel()
+})
+
+test("flows can be cancelled - 2 - finally clauses are run", done => {
+    let steps = 0
+    let finallyHandled = false
+    const start = flow(function*() {
+        steps = 1
+        try {
+            yield Promise.resolve()
+            steps = 2
+        } finally {
+            expect(steps).toBe(1)
+            finallyHandled = true
+        }
+    })
+    const promise = start()
+    promise.then(
+        res => {
+            fail()
+        },
+        err => {
+            expect("" + err).toBe("Error: FLOW_CANCELLED")
+            expect(finallyHandled).toBeTruthy()
+            done()
+        }
+    )
+    promise.cancel()
+})
+
+test("flows can be cancelled - 3 - throw in finally should be catched", done => {
+    const counter = mobx.observable({ counter: 0 })
+    const d = mobx.reaction(() => counter.counter, () => {})
+    mobx.configure({ enforceActions: true })
+
+    const start = flow(function*() {
+        counter.counter = 1
+        try {
+            yield Promise.resolve()
+            counter.counter = 15
+        } finally {
+            counter.counter = 4
+            throw "OOPS"
+        }
+    })
+
+    const promise = start()
+    promise.then(
+        () => fail("flow should not have failed"),
+        err => {
+            expect("" + err).toBe("OOPS")
+            expect(counter.counter).toBe(4)
+            mobx.configure({ enforceActions: false })
+            d()
+            done()
+        }
+    )
+    promise.cancel()
+})
+
+test("flows can be cancelled - 4 - pending Promise will be ignored", done => {
+    let steps = 0
+    const start = flow(function*() {
+        steps = 1
+        yield Promise.reject("This won't be catched anywhere!") // cancel will resolve this flow before this one is throw, so this promise goes uncatched
+        steps = 2
+    })
+
+    const promise = start()
+    promise.then(
+        () => fail(),
+        err => {
+            expect(steps).toBe(1)
+            expect("" + err).toBe("Error: FLOW_CANCELLED")
+            done()
+        }
+    )
+    promise.cancel()
+})
+
+test("flows can be cancelled - 5 - return before cancel", done => {
+    let steps = 0
+    const start = flow(function*() {
+        steps = 1
+        return Promise.resolve(2) // cancel will be to late..
+    })
+
+    const promise = start()
+    promise.then(
+        value => {
+            expect(value).toBe(2), done()
+        },
+        err => {
+            fail()
+        }
+    )
+    promise.cancel() // no-op
+})
+
+test("flows can be cancelled - 5 - flows cancel recursively", done => {
+    let flow1cancelled = false
+    let flow2cancelled = false
+    let stepsReached = 0
+
+    const flow1 = flow(function*() {
+        try {
+            yield Promise.resolve()
+            stepsReached++
+        } finally {
+            flow1cancelled = true
+        }
+    })
+
+    const flow2 = flow(function*() {
+        try {
+            yield flow1()
+            stepsReached++
+        } finally {
+            flow2cancelled = true
+        }
+    })
+
+    const p = flow2()
+    p.then(
+        () => fail(),
+        err => {
+            expect("" + err).toBe("Error: FLOW_CANCELLED")
+            expect(stepsReached).toBe(0)
+            expect(flow2cancelled).toBeTruthy()
+            expect(flow1cancelled).toBeTruthy()
+            done()
+        }
+    )
+    p.cancel()
+})
+
+test("flows yield anything", async () => {
+    let steps = 0
+    const start = flow(function*() {
+        const x = yield 2
+        return x
+    })
+
+    const res = await start()
+    expect(res).toBe(2)
+})